% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AlignmentWindow}
\alias{AlignmentWindow}
\title{AlignmentWindows hold a region's sequence as both a DNAString and a vector of characters}
\usage{
AlignmentWindow(
  Hsapiens_version,
  chrom,
  window_start,
  window_end,
  min_length_for_homopolymer = 5
)
}
\arguments{
\item{Hsapiens_version}{A \pkg{BSgenome} object used to retreive sequence. See \pkg{BSgenome::available.genomes} for choices.}

\item{chrom}{A valid chromosome in Hsapiens_version.}

\item{window_start}{The genomic position on chrom at which to start the region}

\item{window_end}{The genomic position on chrom at which to end the region}

\item{min_length_for_homopolymer}{The minimum number of repeated nucleotides to declare a region a homopolymer. Defaults to 5.
Homopolymers are repeats of a single nucleotide within the reference sequence.
Any single base alteration called by the sequencer that fall within a homopolymer site are likely
to be artifacts, and can be discarded.}
}
\value{
A list with elements start, end, seq, vec. seq is a Biostrings object;
}
\description{
AlignmentWindows hold a region's sequence as both a DNAString and a vector of characters
}
