% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{assess_reversion}
\alias{assess_reversion}
\title{Classify a read to assess the evidence that it contains a reversion}
\usage{
assess_reversion(
  read,
  transcript,
  pathogenic,
  min_nt_qual = 20,
  gr_pathogenic = NULL,
  gr_exclude = NULL
)
}
\arguments{
\item{read}{aardvark::Read object to evaluate}

\item{transcript}{aardvark::TranscriptData object for a gene's transcript}

\item{pathogenic}{aardvark::Mutation object describing pathogenic mutation}

\item{min_nt_qual}{Minimum nucleotide quality score to write read sequence as part of translated sequence.
Reference sequence will be used for nucleotides in the read that are below this quality score.}

\item{gr_pathogenic}{GenomicRanges object spanning the pathogenic mutation}

\item{gr_exclude}{GenomicRanges of regions to exclude from consideration, such as regions
where homopolymers are present. An AlignmentWindow object automatically calculates a
GenomicRangeList of homopolymer regions. Providing this list, or another GenomicRangesList
value to gr_exclude can reduce this source of technical artifacts. Single base deletions
that fall in gr_exclude will be marked read_harbors_variant_in_excluded_region. Larger
alterations are not excluded because they are less likely to be an artifact.}
}
\value{
aardvark::Read object, with updated values for evidence.
The AlignmentWindow object automatically calculates a GenomicRangeList of homopolymer regions. Providing this list, or another
GenomicRangesList value to gr_exclude can reduce this source of technical artifacts.

Possible values for the evidence column:
\itemize{
  \item read_not_informative
  \item read_harbors_variant_in_excluded_region (e.g. a possible homopolymer artifact)
  \item read_harbors_pathogenic_variant_no_reversion
  \item reversion_read_includes_pathogenic_variant
  \item reversion_read_deletion_spans_pathogenic_variant
  \item reversion_read_does_not_include_pathogenic_variant
}
}
\description{
Assesses the evidence that alterations present in a read will result
in a functional trancript when the gene is translated. Sequence alterations are used
to predict the amino acid sequence that would be generated.
}
