% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AW_get_nucleotide}
\alias{AW_get_nucleotide}
\title{helper function to return a specific nucleotide in AlignmentWindow DNAString at pos}
\usage{
AW_get_nucleotide(AW, pos)
}
\arguments{
\item{AW}{Alignment Window object to examine}

\item{pos}{position to return}
}
\value{
character at pos
}
\description{
helper function to return a specific nucleotide in AlignmentWindow DNAString at pos
}
