% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AW_vec}
\alias{AW_vec}
\title{helper function to return a slice of an AlignmentWindow character vector between pos_start and pos_end}
\usage{
AW_vec(AW, pos_start, pos_end)
}
\arguments{
\item{AW}{Alignment Window object to examine}

\item{pos_start}{first position to return}

\item{pos_end}{last position to return}
}
\value{
vector of characters within AW$vec
}
\description{
helper function to return a slice of an AlignmentWindow character vector between pos_start and pos_end
}
