% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{BamData}
\alias{BamData}
\title{Load a BAM file into memory}
\usage{
BamData(fn_bam, chrom, start, end)
}
\arguments{
\item{fn_bam}{path to Bam file}

\item{chrom}{chromosome to read}

\item{start}{read start position}

\item{end}{read end position}
}
\value{
list of reads (data frame) and qualities (list)
}
\description{
This function is a thin wrapper around Rsamtools::scanBam to return reads in a bam file.
You can call scanBam directly if you want different choices for the flags. The flag settings
are:
}
\details{
\itemize{
\item isDuplicate=FALSE
\item isPaired=TRUE
\item isUnmappedQuery=FALSE
\item isSecondaryAlignment=FALSE
\item isSupplementaryAlignment=FALSE
\item isNotPassingQualityControls=FALSE
}
}
