% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{consolidate_ambiguous_candidates}
\alias{consolidate_ambiguous_candidates}
\title{Consolidate duplicate candidate reversions}
\usage{
consolidate_ambiguous_candidates(df_sum, transcript)
}
\arguments{
\item{df_sum}{Read summary generated by aardvark}

\item{transcript}{aardvark::TranscriptData object for a gene's transcript}
}
\description{
Consolidate duplicate candidate reversions
}
