% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{convert_UTF}
\alias{convert_UTF}
\title{convert UTF format to integers, used for sanger conversion}
\usage{
convert_UTF(utf8)
}
\arguments{
\item{utf8}{input string}
}
\value{
converted to integer
}
\description{
convert UTF format to integers, used for sanger conversion
}
