% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{encode_variants}
\alias{encode_variants}
\title{Create a string represetation of all cigar codes that distinguish this from perfect match}
\usage{
encode_variants(rd)
}
\arguments{
\item{rd}{aardvark::Read object}
}
\value{
String representation of the cigar codes (not CIGAR formatted)
}
\description{
Create a string represetation of all cigar codes that distinguish this from perfect match
}
