% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ensembl_cache}
\alias{ensembl_cache}
\title{List of cached transcript data objects}
\format{
a list of TranscriptData objects
}
\usage{
data(ensembl_cache)
}
\description{
List of cached transcript data objects
}
\keyword{ensembl}
