% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ERR2851475}
\alias{ERR2851475}
\title{Example read data ERR2851475}
\format{
a list of BAM data elements
}
\usage{
data(ERR2851475)
}
\description{
Example read data ERR2851475
}
\keyword{ERR2851475}
