% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{genomicRangesFromMutation}
\alias{genomicRangesFromMutation}
\title{Helper function to convert a mutation to a GenomicRanges object}
\usage{
genomicRangesFromMutation(germline_mutation)
}
\arguments{
\item{germline_mutation}{mutation object}
}
\value{
a GenomicRanges object
}
\description{
Helper function to convert a mutation to a GenomicRanges object
}
