% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.split.col}
\alias{get.split.col}
\title{split vector all at once}
\usage{
get.split.col(v, string, col = 0, last = F, first = F)
}
\arguments{
\item{v}{vector}

\item{string, }{split value}

\item{col}{column to return}

\item{last}{return last}

\item{first}{return first}
}
\description{
split vector all at once
}
