% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hg19_ensembl_coding_transcripts}
\alias{hg19_ensembl_coding_transcripts}
\title{List of cached biomaRt::getSequence results
Note there is no MANE data for the hg19 build}
\format{
a list of biomaRt::getSequence results
}
\usage{
data(hg19_ensembl_coding_transcripts)
}
\description{
List of cached biomaRt::getSequence results
Note there is no MANE data for the hg19 build
}
\keyword{ensembl}
