% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hg38_ensembl_coding_transcripts}
\alias{hg38_ensembl_coding_transcripts}
\title{List of cached biomaRt::getSequence results}
\format{
a list of biomaRt::getSequence results
}
\usage{
data(hg38_ensembl_coding_transcripts)
}
\description{
List of cached biomaRt::getSequence results
}
\keyword{ensembl}
