% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{hsh_get}
\alias{hsh_get}
\title{Return value of key in hash H}
\usage{
hsh_get(H, key, na.if.not.found = FALSE)
}
\arguments{
\item{H}{hash object}

\item{key}{key to set}

\item{na.if.not.found}{If key not in H and TRUE, return NA; if FALSE, raise an error}
}
\description{
Return value of key in hash H
}
