% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{hsh_set}
\alias{hsh_set}
\title{Set key to value in hash H}
\usage{
hsh_set(H, key, value)
}
\arguments{
\item{H}{hash object}

\item{key}{key to set}

\item{value}{new value of key in H}
}
\description{
Set key to value in hash H
}
