% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{locally_realign_read}
\alias{locally_realign_read}
\title{Local realignment of read ends to improve alignment near a germline mutation}
\usage{
locally_realign_read(
  read,
  gr_mut,
  align_window,
  near_bound = 10,
  min_nt_qual = 20,
  allow_insertions_in_realign = FALSE,
  realign_side = "both"
)
}
\arguments{
\item{read}{Read to modify}

\item{gr_mut}{GenomicRanges location of mutation}

\item{align_window}{AlignmentWindow object that holds reference sequence search space for realignment}

\item{near_bound}{number of nucleotides adjacent to a germline mutation to consider off of the end of a read for local realignment}

\item{min_nt_qual}{minimal nucleotide quality score to count when evaluating perfect matches in a candidate alignment}

\item{allow_insertions_in_realign}{whether to allow insertions in the realignment}

\item{realign_side}{which side to attempt realignment, one of "left", "right", "both"}
}
\value{
returns the updated read
}
\description{
Primarily effective when the read ends at or just over the germline mutation
if both left and right side are softclipped, need to consider them separately;
if neither is softclipped can consider both sides.
}
