% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{match.idx}
\alias{match.idx}
\title{return indices into A and B restricted to perfect matches where idx.A[i] == idx.B[i] for each i in matched pairs}
\usage{
match.idx(A, B, allow.multiple.B = FALSE)
}
\arguments{
\item{A}{vector}

\item{B}{vector}

\item{allow.multiple.B}{If true, A->B can be one to many; otherwise return first match}
}
\description{
return indices into A and B restricted to perfect matches where idx.A[i] == idx.B[i] for each i in matched pairs
}
