% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{merge_softclipped_into_read}
\alias{merge_softclipped_into_read}
\title{Attempt to merge the softclipped portion of a read segment into the segment for realignment}
\usage{
merge_softclipped_into_read(read, idx_from, idx_into)
}
\arguments{
\item{read}{aardvark::Read object containing segments to test}

\item{idx_from}{index of cigar_ranges that is the source of clipping}

\item{idx_into}{index of cigar_ranges that is the target of merge}
}
\value{
Returns the updated read object
}
\description{
This merge is only saved as the final aligned read if the new alignment is an improvement.
}
