% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Mutation}
\alias{Mutation}
\title{Generate an object that holds information about a mutation.}
\usage{
Mutation(chrom, pos, seq_ref, seq_alt)
}
\arguments{
\item{chrom}{chromosome}

\item{pos}{start position of the mutation in on chromsosome chrom}

\item{seq_ref}{character string indicating the reference sequence}

\item{seq_alt}{character string indicating the mutation sequence}
}
\value{
a list with elements that describe the consequences of the mutation
}
\description{
seq_ref and seq_alt are strings that should hold either a dash or a nucleotide.
To indicate a two nucleotide deletion, where the reference sequence that was deleted is AC,
pass "AC" for seq_ref and "--" for seq_alt To indicate a one nucleotide insertion, where
the mutation inserts a T, pass "-" for seq_ref and "T" for seq_alt To indicate a single
nucleotide variation where the reference sequence was a "C" and the mutation converts the C to
an, pass "C" for seq_ref and "A" for seq_alt
}
