% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_reversion_frequency}
\alias{plot_reversion_frequency}
\title{Barplot the frequency of each reversion within a sample}
\usage{
plot_reversion_frequency(
  sums,
  height = 1,
  xinterval = 10,
  exclude_blacklist = TRUE,
  min_freq = 2
)
}
\arguments{
\item{sums}{Summary of reversions}

\item{height}{height of each bar}

\item{xinterval}{interval of x axis ticks}

\item{exclude_blacklist}{exclude candidate reversions that intersect a blacklist region}

\item{min_freq}{minimum number of observations to plot a candidate, default 2}
}
\description{
Barplot the frequency of each reversion within a sample
}
