% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_reversion_summary}
\alias{plot_reversion_summary}
\title{Plot a reversion map showing the physical locations of the reversions}
\usage{
plot_reversion_summary(
  reversion_summary,
  genome_version = 38,
  hsapiens_object,
  biomart_object,
  pos_start,
  pos_end,
  exclude_blacklist = TRUE,
  min_freq = 2,
  text_size_adjustment = 1
)
}
\arguments{
\item{reversion_summary}{AARDVARK reversion summary, either loaded from the
text file generated by \code{\link[=write_read_summary]{write_read_summary()}}
or the summary dataframe returned by a call to \code{\link[=summarize_candidates]{summarize_candidates()}}}

\item{genome_version}{either 38 or 19}

\item{hsapiens_object}{hsapiens object from either \link{BSgenome.Hsapiens.UCSC.hg38}
or BSgenome.Hsapiens.UCSC.hg19}

\item{biomart_object}{a BioMart object with a genome version matching the
genome_version parameter, created by a call to the \code{\link[biomaRt:useEnsembl]{biomaRt::useEnsembl()}} function}

\item{pos_start}{leftmost genomic position to plot}

\item{pos_end}{rightmost genomic position to plot}

\item{exclude_blacklist}{boolean, if TRUE don't plot reversions marked with the evidence "overlaps_blacklist"; defaults TRUE}

\item{min_freq}{minimum number of observations required to plot a candidate reversion; defaults 2}

\item{text_size_adjustment}{adjustment to text size; larger values give larger text. Defaults 1}
}
\description{
Plot a reversion map showing the physical locations of the reversions
}
