% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_reversion_summary_from_file}
\alias{plot_reversion_summary_from_file}
\title{Plot a reversion map showing the physical locations of the reversions in summary sums,
'reading in reversion data from a summary generated by AARDVARK}
\usage{
plot_reversion_summary_from_file(
  filename,
  genome_version = 38,
  hsapiens_object,
  biomart_object,
  pos_start,
  pos_end,
  exclude_blacklist = TRUE,
  min_freq = 2
)
}
\arguments{
\item{filename}{AARDVARK summary file to read}

\item{genome_version}{either 38 or 19}

\item{hsapiens_object}{hsapiens object from either BSgenome.Hsapiens.UCSC.hg38 or BSgenome.Hsapiens.UCSC.hg19}

\item{biomart_object}{a biomart object matching the genome_version parameter, created with the biomaRt package}

\item{pos_start}{leftmost position for plot}

\item{pos_end}{rightmost position for plot}

\item{exclude_blacklist}{boolean, if TRUE don't plot reversions marked with the evidence "overlaps_blacklist"; defaults TRUE}

\item{min_freq}{minimum number of observations to plot a candidate; defaults 2}
}
\description{
Plot a reversion map showing the physical locations of the reversions in summary sums,
'reading in reversion data from a summary generated by AARDVARK
}
