% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Read}
\alias{Read}
\title{Generate an object that holds information about a read}
\usage{
Read(qname, cigar, chrom, pos, seq, qual)
}
\arguments{
\item{qname}{read name, usually assigned by the sequencer}

\item{cigar}{CIGAR string}

\item{chrom}{chromosome}

\item{pos}{read start position, first non-clipped nucleotide}

\item{seq}{DNAString object that represents the nucleotide sequence as}

\item{qual}{vector of quality scores for nucleotides in seq; must have same length as seq.}
}
\value{
a list with elements that describe the read
}
\description{
The Read object can be decorated with additional information by
generated through local realignment.
}
\seealso{
\code{\link{read_from_BamData}}

\code{\link{assess_reversion}}

\code{\link{realign_read}}
}
