% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{read_from_BamData}
\alias{read_from_BamData}
\title{Instantiate a Read from a list of reads}
\usage{
read_from_BamData(bamdata, read_index)
}
\arguments{
\item{bamdata}{BamData object}

\item{read_index}{the index of the read to return.}
}
\value{
An aardvark::read object
}
\description{
This function consumes a BamData object and the index of a read within that object and
generates a single aardvark::Read object.
}
