% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{read_to_genome_sequence}
\alias{read_to_genome_sequence}
\title{Given a read, return a simple data frame that write the reference sequence
where there is a perfect match and D or I for deletion/insertion.}
\usage{
read_to_genome_sequence(read, alignment_window)
}
\arguments{
\item{read}{aardvark::Read object under consideration}

\item{alignment_window}{aardvark::AlignmentWindow object that holds reference sequence search space for realignment}
}
\description{
Given a read, return a simple data frame that write the reference sequence
where there is a perfect match and D or I for deletion/insertion.
}
