% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{realign_repeat_pathogenic_deletions}
\alias{realign_repeat_pathogenic_deletions}
\title{If a pathogenic mutation is a deletion located adjacent to a DNA sequence
that repeats the deleted nucleotides, there will likely be alignments
that put the deletion on the non-germline deleted repeat segment intead of
the germline deletion. This looks reasonable to the aligner and can frequently
result in a higher alignment (e.g. because it makes a contiguous deletion)
but we know this is extremely implausible due to the position of the germline
deletion. This function looks for exactly that situation and remedies it.}
\usage{
realign_repeat_pathogenic_deletions(
  read,
  gr_pathogenic,
  pathogenic_mutation,
  align_window
)
}
\arguments{
\item{read}{aardvark::Read object under consideration}

\item{gr_pathogenic}{GenomicRanges spanning the pathogenic mutation}

\item{pathogenic_mutation}{the pathogenic mutation}

\item{align_window}{aardvark::AlignmentWindow object that holds reference sequence search space for realignment}
}
\description{
If a pathogenic mutation is a deletion located adjacent to a DNA sequence
that repeats the deleted nucleotides, there will likely be alignments
that put the deletion on the non-germline deleted repeat segment intead of
the germline deletion. This looks reasonable to the aligner and can frequently
result in a higher alignment (e.g. because it makes a contiguous deletion)
but we know this is extremely implausible due to the position of the germline
deletion. This function looks for exactly that situation and remedies it.
}
