% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{realign_sequence_in_region}
\alias{realign_sequence_in_region}
\title{Align a nucleotide sequence in a region of the reference}
\usage{
realign_sequence_in_region(ref_pos, ref_seq, query_str)
}
\arguments{
\item{ref_pos}{chromosome position in genome at which original alignment started}

\item{ref_seq}{DNAString of nucleotides in reference sequence}

\item{query_str}{character string of nucleotides to align in ref_seq}
}
\value{
list with length of longest alignment, start and end in reference and sequence space, and the alignment object
}
\description{
performs a Biostrings::pairwise alignment of query_str and ref_seq
and identifies the longest perfet alignment between the two sequences
}
