% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{realign_to_ref_with_pathogenic_deletion}
\alias{realign_to_ref_with_pathogenic_deletion}
\title{Assess whether there is a perfect match between aligned and reference that is missing the pathogenic deletion}
\usage{
realign_to_ref_with_pathogenic_deletion(
  read,
  pathogenic,
  gr_pathogenic,
  align_window
)
}
\arguments{
\item{read}{Read to modify}

\item{pathogenic}{aardvark::Mutation object describing pathogenic mutation}

\item{gr_pathogenic}{GenomicRanges location of mutation}

\item{align_window}{AlignmentWindow object that holds reference sequence search space for realignment}
}
\value{
returns TRUE/FALSE
}
\description{
Assess whether there is a perfect match between aligned and reference that is missing the pathogenic deletion
}
