% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove_row}
\alias{remove_row}
\title{Remove an arbitrary row from a dataframe and return the result}
\usage{
remove_row(df, row_id_to_remove)
}
\arguments{
\item{df}{dataframe to modify}

\item{row_id_to_remove}{row to remove}
}
\description{
Remove an arbitrary row from a dataframe and return the result
}
