% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{replace_row}
\alias{replace_row}
\title{replace an existing row with the contents of df_new (may be multiple rows)}
\usage{
replace_row(df, row_id_to_replace, df_new)
}
\arguments{
\item{df}{dataframe to modify}

\item{row_id_to_replace}{row to replace}

\item{df_new}{new dataframe, columns must match df}
}
\description{
replace an existing row with the contents of df_new (may be multiple rows)
}
