% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Returns reverse complement of a vector of A,C,G,T}
\usage{
reverse_complement(ts_fwd)
}
\arguments{
\item{ts_fwd}{character string of transcript to generate the reverse complement}
}
\value{
reverse complement vector of ts_fwd
}
\description{
Returns reverse complement of a vector of A,C,G,T
}
