% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{transcript_ATM}
\alias{transcript_ATM}
\title{Transcript data for ATM}
\format{
a TranscriptData object
}
\usage{
data(transcript_ATM)
}
\description{
Transcript data for ATM
}
\keyword{ATM}
