% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{transcript_BRCA1}
\alias{transcript_BRCA1}
\title{Transcript data for BRCA1, HG38}
\format{
a TranscriptData object
}
\usage{
data(transcript_BRCA1)
}
\description{
Transcript data for BRCA1, HG38
}
\keyword{BRCA1}
