% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{transcript_cache_hg38}
\alias{transcript_cache_hg38}
\title{List of cached transcript data objects}
\format{
a list of transcript objects
}
\usage{
data(transcript_cache_hg38)
}
\description{
List of cached transcript data objects
}
\keyword{ensembl}
