% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TranscriptData}
\alias{TranscriptData}
\title{Generate an object that holds genome location information about transcript}
\usage{
TranscriptData(biomart_object, ensdb_object, transcript_id)
}
\arguments{
\item{biomart_object}{instantated biomart object}

\item{ensdb_object}{instantiated ensemblDB object}

\item{transcript_id}{ENBEMBL identity of a transcript to look up using biomaRt}
}
\value{
a list with elements that describe the transcript: transcript sequence as a DNAString object, the strand (+ or -) and a data frame of the nucleotides listing pos, seq (nucleotide), exon_id, strand, is_splice (whether the position is a splice junction)
}
\description{
This function uses biomaRt to fetch relevant gene model data about a transcript.
}
