% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{translate_cigar}
\alias{translate_cigar}
\title{Generate the amino acid sequence by applying the alterations present in a read to a given transcript}
\usage{
translate_cigar(transcript, read, pathogenic = NULL, min_nt_qual)
}
\arguments{
\item{transcript}{an aardvark::TranscriptData object for affected gene}

\item{read}{aardvark::Read object to evaluate}

\item{pathogenic}{aardvark::Mutation object describing a pathogenic mutation (optional)}

\item{min_nt_qual}{Minimum nucleotide quality score to write read sequence as part of translated sequence.
Reference sequence will be used for nucleotides in the read that are below this quality score.}
}
\value{
A Biostrings::AAString object representing the amino acid sequence of complete transcript; * is STOP.
}
\description{
If pathogenic is not passed, the native sequence modified by the value of read is returned.
If both read and pathogenic are passed, the impact of both alterations can affect the amino
acid sequence. Alterations obsered in the read supercede alterations in pathogenic if the read overlaps
the pathogenic alteration.
}
\details{
If the read cannot be evaluated (e.g. because it contains a nucleotide coded as N)
translate_cigar returns NULL
}
