% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aardvark.R
\name{write_read_summary}
\alias{write_read_summary}
\title{Write out a summary of all reads that support these results}
\usage{
write_read_summary(
  read_summary,
  pathogenic_mutation,
  path_read_summary,
  sample_id,
  transcript_id
)
}
\arguments{
\item{read_summary}{Read summary generated by aardvark}

\item{pathogenic_mutation}{aardvark::Mutation object describing pathogenic mutation}

\item{path_read_summary}{path to file to write}

\item{sample_id}{Sample ID to write within summary file}

\item{transcript_id}{Ensembl identifier of the gene transcript to which reads were compared}
}
\description{
Write out a summary of all reads that support these results
}
