% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exon_cache_hg38}
\alias{exon_cache_hg38}
\title{List of cached exon data objects}
\format{
a list of ensembl exon objects
}
\usage{
data(exon_cache_hg38)
}
\description{
List of cached exon data objects
}
\keyword{ensembl}
