% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecasters.R
\name{get_forecasters}
\alias{get_forecasters}
\title{Return the desired forecaster function}
\usage{
get_forecasters(signals, ahead)
}
\arguments{
\item{signals}{Tibble with columns \code{data_source}, \code{signal}, 
\code{start_day} that specify which variables from the COVIDcast API will 
be used by forecaster. Each row of signals represents a separate signal, 
and first row is taken to be the response.}

\item{ahead}{The number of incidence periods ahead to forecast the response.
For \code{incidence_period = "epiweek"}, one of 1, 2, 3, 4.}
}
\value{
A list with an element named \code{aardvark_forecaster}, 
    which is itself a list consisting of the forecaster function and a \code{type} 
    string (one of \code{c("standalone","ensemble")}), with \code{type = "ensemble"} 
    now deprecated. Unavailable forecasters are marked as 
    \code{list(forecaster = NA, type = "standalone")}.
}
\description{
The \link[evalcast]{evalcast-package} production 
    evaluator will first call this function to determine all the forecasters
    available for the given parameter specifications. It expects to get back 
    a named list of lists of forecaster functions and types. If a forecaster 
    function is not available for a given set of parameters, an 
    \code{NA} should returned instead of a function. This tells the 
    evaluator to ignore that forecaster in that run.
}
\examples{
    signals <- dplyr::tibble(data_source = "jhu-csse",
    signal = c("deaths_incidence_num", "confirmed_incidence_num"), start_day = "2020-03-07")
    ahead <- 1
    aardvark_forecaster <- aardvark::get_forecasters(signals = signals, ahead = ahead)[[1]]$forecaster
}
