% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecasters.R
\name{get_forecasters}
\alias{get_forecasters}
\title{Get the Aardvark state death production forecaster}
\usage{
get_forecasters(
  response = "jhu-csse_deaths_incidence_num",
  incidence_period = c("epiweek"),
  ahead,
  forecast_date,
  geo_type = c("state", "county", "hrr", "msa"),
  n_locations = 52
)
}
\arguments{
\item{response}{The response variable (for this forecaster, should be 
"jhu-csse_deaths_incidence_num" -- the incident deaths reported by Johns Hopkins 
University).}

\item{incidence_period}{The incidence period. This state death forecaster
currently only supports forecasts at the "epiweek" level.}

\item{ahead}{The number of incidence periods ahead to forecast state deaths.
One of 1, 2, 3, 4.}

\item{forecast_date}{The date of the forecast.}

\item{geo_type}{the geographic type (e.g "state" or "county" or
"hrr" or "msa"... but for now only the first two),}

\item{n_locations}{The number of geo locations. Up to 52 for \code{geo_type = "state"}
(states + D.C. and Puerto Rico)}
}
\value{
A list with an element named \code{aardvark_state_death_forecaster}, 
    which is itself a list consisting of the forecaster function and a \code{type} 
    string (one of \code{c("standalone","ensemble")}), with \code{type = "ensemble"} now 
    deprecated. Unavailable forecasters are marked as 
    \code{list(forecaster = NA, type = "standalone")}.
}
\description{
The \link[evalcast]{evalcast-package} production 
    evaluator will first call this function to determine all the forecasters
    available for the given parameter specifications. It expects to get back 
    a named list of lists of forecasting functions and types. If a forecasting 
    function is not available for a given set of parameters, an 
    \code{NA} should returned instead of a function. This tells the 
    evaluator to ignore that forecaster in that run.
}
\examples{
# Return the Aardvark state death forecaster function for 1 epiweek ahead
library(aardvark)

ahead = 1
my_forecaster <- get_forecasters(response = "jhu-csse_deaths_incidence_num",
                                 ahead = ahead)[["aardvark_state_death_forecaster"]][["forecaster"]]
}
