% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_reversion_summary}
\alias{plot_reversion_summary}
\title{Plot a reversion map showing the physical locations of the reversions in summary sums}
\usage{
plot_reversion_summary(
  sums,
  xlim,
  height = 1,
  xinterval = 250,
  exclude_blacklist = TRUE,
  min_freq = 2
)
}
\arguments{
\item{sums}{Summary of reversions}

\item{xlim}{x axis limits}

\item{height}{height of a single row; defaults 1}

\item{xinterval}{X axis interval between ticks in the label; defaults 250}

\item{exclude_blacklist}{boolean, if TRUE don't plot reversions marked with the evidence "overlaps_blacklist"; defaults TRUE}

\item{min_freq}{minimum number of observations to plot a candidate; defaults 2}
}
\description{
Plot a reversion map showing the physical locations of the reversions in summary sums
}
