% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sequence_cache_hg38}
\alias{sequence_cache_hg38}
\title{List of cached transcript data objects}
\format{
a list of sequence objects
}
\usage{
data(ensembl_cache)
}
\description{
List of cached transcript data objects
}
\keyword{ensembl}
