% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{transcript_BRCA2}
\alias{transcript_BRCA2}
\title{Transcript data for BRCA2}
\format{
a TranscriptData object
}
\usage{
data(transcript_BRCA2)
}
\description{
Transcript data for BRCA2
}
\keyword{BRCA2}
